+++ C/src/utils.h
@@ -23,6 +23,6 @@ void PrintTimes(const std::vector<TimePoint>& timestamps);
 void RestartListener(Session& sess);
 void ClearQueue(std::queue<vector<uint8_t>>& q);
 bool WithProbability(double probability);
-void WritePulseAmplitudes(const std::vector<double>& click_amps, const std::vector<TimePoint>& timestamps, const std::string& filename);
-void WriteArray(const Eigen::VectorXd& array, const std::vector<TimePoint>& timestamps, const std::string& filename);
+void WritePulseAmplitudes(const std::vector<float>& click_amps, const std::vector<TimePoint>& timestamps, const std::string& filename);
+void WriteArray(const Eigen::VectorXf& array, const std::vector<TimePoint>& timestamps, const std::string& filename);
 void WriteDataToCerr(vector<TimePoint>& dataTimes, vector<vector<uint8_t>>& dataBytesSaved);
